+++ src/mgmt/api/mod.rs
@@ -5,6 +5,7 @@ pub mod arrays;
 pub mod volumes;
 pub mod exports;
 pub mod fstemplates;
+pub mod images;
 pub mod pallets;
 pub mod slabs;
 pub mod discovery;
@@ -38,6 +39,7 @@ pub fn router(state: Arc<AppState>) -> Router {
         .nest("/api/v1/exports", exports::router(state.clone()))
         .nest("/api/v1/slabs", slabs::router(state.clone()))
         .nest("/api/v1/pallets", pallets::router(state.clone()))
+        .nest("/api/v1/images", images::router(state.clone()))
         .nest("/api/v1/fstemplates", fstemplates::router(state.clone()))
         .nest("/api/v1/moves", moves::router(state.clone()))
         .nest("/api/v1/discovery", discovery::router(state.clone()))